requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pandas>=2.1.0
openpyxl>=3.1.2
pyyaml>=6.0.1
python-multipart>=0.0.6
beautifulsoup4>=4.12.0
lxml>=4.9.0
